修改生成工作进程方式及信号处理

+++ include/conf.h
@@ -21,7 +21,7 @@
 
 typedef struct s_conf_t {
     int m_daemon;
-    int m_worker_processes;
+    intptr_t m_worker_processes;
     int m_max_connections;
     int m_timer_resolution;
     int m_connection_timeout;
